rename: 목데이터 파일명 변경 issue: SCRUM-123
